Add Click event to change Pages
Add click event to the divs in the header to change from page
to page in index.js.

Add function ChangeHeader to header.js, to change the header
style of the divs so it marks which one is selected.

Add function erasePage to erase the page that is inside a
main element.